rename multi_pose_manager.py to octree_manager.py
