+++ octreelib/octree_manager/__init__.py
@@ -1,5 +1,5 @@
-import octreelib.octree_manager.multi_pose_manager as multi_pose_manager_module
+import octreelib.octree_manager.octree_manager as octree_manager_module
 
-from octreelib.octree_manager.multi_pose_manager import *
+from octreelib.octree_manager.octree_manager import *
 
-__all__ = multi_pose_manager.__all__
+__all__ = octree_manager_module.__all__
